modified:   .env 	modified:   src/main.js 	modified:   src/router/index.js 	modified:   src/shared/Login.vue 	modified:   src/shared/Navbar.vue 	modified:   src/shared/Sidebar.vue 	new file:   src/views/Success.vue
