+++ .env
@@ -1,0 +1,8 @@
+# FIREBASE CONFIG
+VITE_API_KEY="AIzaSyAPO7PfInLyN3wKPviHaf9hdd9yc5Qr7is";
+VITE_AUTH_DOMAIN="luminescence-ckms39th.firebaseapp.com";
+VITE_PROJECT_ID="luminescence-ckms39th";
+VITE_STORAGE_BUCKET="luminescence-ckms39th.appspot.com";
+VITE_MESSAGING_SENDER_ID="882298335163";
+VITE_APP_ID="1:882298335163:web:55e8aac9538de0551b5e9e";
+VITE_MEASUREMENT_ID="G-RMVG3ZBN6E";
