+++ src/shared/Login.vue
@@ -1,0 +1,222 @@
+<template>
+  <button class="login-button" @click="showLogin">Login</button>
+  <div class="login-overlay" v-if="showingLogin">
+    <div class="login-container">
+      <h2>Login</h2>
+      <form>
+        <div class="form-group">
+          <input v-model="email" type="email" placeholder="Email" />
+        </div>
+        <div class="form-group">
+          <input
+              v-model="password"
+              type="password"
+              placeholder="Password"
+          />
+        </div>
+        <button type="button" @click="login" :disabled="loggingIn">
+          {{ loggingIn ? "Logging in..." : "Log in" }}
+        </button>
+      </form>
+      <button class="close-button" @click="hideLogin">Cancel</button>
+    </div>
+  </div>
+</template>
+
+<script setup>
+import {ref} from "vue";
+
+const email = ref("");
+const password = ref("");
+const loggingIn = ref(false);
+const showingLogin = ref(false);
+
+const correctEmail = "123@email.com";
+const correctPasswd = "1234";
+
+async function login() {
+  loggingIn.value = true;
+  try {
+    if (email.value === correctEmail && password.value === correctPasswd) {
+      console.log("true111");
+      window.location.href = "/#/success";
+    } else {
+      console.log("false111");
+    }
+  } catch (error) {
+    console.error(error);
+  }
+  loggingIn.value = false;
+}
+
+function showLogin() {
+  showingLogin.value = true;
+}
+
+function hideLogin() {
+  showingLogin.value = false;
+}
+</script>
+
+<style scoped>
+
+.login-button {
+  background-color: transparent;
+  color: white;
+  cursor: pointer;
+  font-family: 'Recursive', sans-serif;
+  font-weight: bold;
+  border-radius: 0.45em;
+  text-align: center;
+  padding-top: 15px;
+  padding-bottom: 15px;
+}
+
+.login-button:hover {
+  border-color: hsl(104, 100%, 86%);
+  border-radius: 0.45em;
+  background-color: hsl(104, 100%, 86%);
+  box-shadow: 0px 0px 0.75em 1px hsl(104, 100%, 86%);
+  animation: text-flicker 6s linear infinite;
+  color: black;
+}
+
+.login-overlay {
+  position: fixed;
+  top: 0;
+  left: 0;
+  width: 100%;
+  height: 100%;
+  background-color: rgba(0, 0, 0, 0.7);
+  z-index: 9999;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  animation: fade-in-animation ease-in 0.2s;
+}
+
+.login-container {
+  background-color: #fff;
+  width: 280px;
+  padding: 30px;
+  border-radius: 10px;
+  box-shadow: 0 0 10px rgba(0, 0, 0, 0.3);
+  text-align: center;
+  transform: translateY(0%);
+  transition: transform 0.5s ease-in-out;
+}
+
+.login-container.show {
+  transform: translateY(0);
+}
+
+.login-container h2 {
+  margin-bottom: 30px;
+  font-size: 24px;
+  font-weight: bold;
+  color: #333;
+  text-transform: uppercase;
+  letter-spacing: 2px;
+}
+
+.login-container form {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+}
+
+.login-container form .form-group {
+  margin-bottom: 20px;
+  width: 100%;
+}
+
+.login-container form label {
+  display: block;
+  margin-bottom: 5px;
+  font-size: 14px;
+  color: #555;
+}
+
+.login-container form input {
+  display: block;
+  width: 100%;
+  margin-bottom: 10px;
+  padding: 10px;
+  border-radius: 5px;
+  border: none;
+  background-color: rgba(0, 105, 217, 0.7);
+  transition: background-color 0.5s ease-in-out;
+}
+
+.login-container form input:focus {
+  outline: none;
+  background-color: #0069d9;
+}
+
+.login-container form button {
+  display: block;
+  width: 100%;
+  padding: 10px;
+  background-color: #007bff;
+  color: #fff;
+  border: none;
+  border-radius: 5px;
+  cursor: pointer;
+  transition: background-color 0.5s ease-in-out;
+}
+
+.login-container form button:hover {
+  background-color: #0069d9;
+}
+
+.login-container form button:active {
+  transform: scale(0.95);
+}
+
+.close-button {
+  background-color: transparent;
+  border: none;
+  color: #888;
+  font-size: 16px;
+  cursor: pointer;
+  margin-top: 20px;
+}
+
+
+@media (max-width: 2000px) {
+  .login-button{
+    font-size: 1.1rem;
+    width: 100px;
+  }
+}
+
+@media (max-width: 1100px) {
+  .login-button{
+    font-size: 0.9rem;
+    width: 70px;
+  }
+}
+
+@media (max-width: 800px) {
+  .login-button{
+    margin-right: 10px;
+  }
+}
+
+@media (max-width: 480px){
+  .login-button{
+    margin-right: 20px;
+    padding-left: 5px;
+    padding-right: 5px;
+  }
+}
+
+@media (max-width: 380px){
+  .login-button{
+    margin-right: 30px;
+    padding-left: 3px;
+    padding-right: 3px;
+    width: 100%;
+  }
+}
+</style>
